Change mustache file to use httprb. Regeneragte code.
